# 8360559: Optimize Math.sinh for x86 64 bit platforms