--- a/src/hotspot/share/interpreter/abstractInterpreter.cpp
+++ b/src/hotspot/share/interpreter/abstractInterpreter.cpp
@@ -137,6 +137,7 @@ AbstractInterpreter::MethodKind AbstractInterpreter::method_kind(const methodHan
       case vmIntrinsics::_dsin:              return java_lang_math_sin;
       case vmIntrinsics::_dcos:              return java_lang_math_cos;
       case vmIntrinsics::_dtan:              return java_lang_math_tan;
+      case vmIntrinsics::_dsinh:             return java_lang_math_sinh;
       case vmIntrinsics::_dtanh:             return java_lang_math_tanh;
       case vmIntrinsics::_dcbrt:             return java_lang_math_cbrt;
       case vmIntrinsics::_dabs:              return java_lang_math_abs;
@@ -199,6 +200,7 @@ vmIntrinsics::ID AbstractInterpreter::method_intrinsic(MethodKind kind) {
   case java_lang_math_sin         : return vmIntrinsics::_dsin;
   case java_lang_math_cos         : return vmIntrinsics::_dcos;
   case java_lang_math_tan         : return vmIntrinsics::_dtan;
+  case java_lang_math_sinh        : return vmIntrinsics::_dsinh;
   case java_lang_math_tanh        : return vmIntrinsics::_dtanh;
   case java_lang_math_cbrt        : return vmIntrinsics::_dcbrt;
   case java_lang_math_abs         : return vmIntrinsics::_dabs;
@@ -304,6 +306,7 @@ void AbstractInterpreter::print_method_kind(MethodKind kind) {
     case java_lang_math_sin     : tty->print("java_lang_math_sin"     ); break;
     case java_lang_math_cos     : tty->print("java_lang_math_cos"     ); break;
     case java_lang_math_tan     : tty->print("java_lang_math_tan"     ); break;
+    case java_lang_math_sinh    : tty->print("java_lang_math_sinh"    ); break;
     case java_lang_math_tanh    : tty->print("java_lang_math_tanh"    ); break;
     case java_lang_math_cbrt    : tty->print("java_lang_math_cbrt"    ); break;
     case java_lang_math_abs     : tty->print("java_lang_math_abs"     ); break;