@jddarcy 2025-07-07
Are there plans for a cosh intrinsic as follow-up work? There was a previous effort to intrinsify tanh.